+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added the option for Layers on the Analysis Page Map
   [GEOD-1225](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1225)
 
+### Changed
+- The Search Page Filter Icons reflect the dropdown selection
+  [GEOD-1224](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1224)
+
 ## [3.0.0-beta.4] - 2019-05-06
 
 ### Added